Refactor attn patcher
