+++ src/comfyui_daam/daam/__init__.py
@@ -1,26 +1,21 @@
 from __future__ import annotations
 from .analyzer import PromptAnalyzer
 from .heatmap import GlobalHeatMap, HeatMapProcessor
-from .patcher import BaseAttentionPatcher, CrossAttentionPatcher
 from .util import (
     UNetCrossAttentionLocator,
     is_output_connected,
     ObjectHooker,
     AggregateHooker,
 )
-from .flux_patcher import FluxAttentionPatcher
 from .attention import attention_sub_quad_patched
 
 __all__ = [
     "GlobalHeatMap",
     "HeatMapProcessor",
-    "CrossAttentionPatcher",
     "UNetCrossAttentionLocator",
     "PromptAnalyzer",
     "is_output_connected",
     "ObjectHooker",
     "AggregateHooker",
-    "BaseAttentionPatcher",
-    "FluxAttentionPatcher",
     "attention_sub_quad_patched",
 ]
